+++ test/hexapdf/type/acro_form/test_field.rb
@@ -155,6 +155,11 @@ describe HexaPDF::Type::AcroForm::Field do
       assert_equal(5, widget[:X])
     end
 
+    it "sets the print flag on the widget" do
+      widget = @field.create_widget(@page, X: 5)
+      assert_equal([:print], widget.flags)
+    end
+
     it "adds the new widget to the given page's annotations" do
       widget = @field.create_widget(@page)
       assert_equal([widget], @page[:Annots].value)
@@ -179,7 +184,7 @@ describe HexaPDF::Type::AcroForm::Field do
       refute_same(widget1, kids[0])
       assert_same(widget2, kids[1])
       assert_nil(@field[:Rect])
-      assert_equal({Rect: [1, 2, 3, 4], Type: :Annot, Subtype: :Widget, Parent: @field},
+      assert_equal({Rect: [1, 2, 3, 4], Type: :Annot, Subtype: :Widget, Parent: @field, F: 4},
                    kids[0].value)
       assert_equal([2, 1, 4, 3], kids[1][:Rect].value)